Improved widget class hierarchy
Created a Frame class and a Layout class. The class members and functions are distributed to these classes accordingly. Now frames do not have member variables for the current mouse state they are in. Only widgets do. Layouts are frames with multiple child frames. The window does not have a widget, but a layout member.